requirements.txt
appdirs==1.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
docker-pycreds==0.4.0
einops==0.6.1
filelock==3.12.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
Jinja2==3.1.2
lit==16.0.3
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
pathtools==0.1.2
protobuf==4.23.0
psutil==5.9.5
PyYAML==6.0
requests==2.30.0
sentry-sdk==1.22.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
torch==2.0.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.15.2
